+++ converse.js
@@ -2358,6 +2358,7 @@
                 'click .toggle-smiley': 'toggleEmoticonMenu',
                 'click .toggle-smiley ul li': 'insertEmoticon',
                 'click .toggle-clear': 'clearChatRoomMessages',
+                'click .toggle-call': 'toggleCall',
                 'click .toggle-participants a': 'toggleOccupants',
                 'keypress textarea.chat-textarea': 'keyPressed',
                 'mousedown .dragresize-tm': 'onDragResizeStart'
@@ -2427,7 +2428,15 @@
                 this.toggleOccupants();
                 return this;
             },
-
+			
+            toggleCall: function (ev) {
+                ev.stopPropagation();
+                converse.emit('callButtonClicked', {
+                    connection: converse.connection,
+                    model: this.model
+                });
+            },
+			
             toggleOccupants: function (ev) {
                 if (ev) {
                     ev.preventDefault();