+++ docs/CHANGES.rst
@@ -15,6 +15,8 @@ Changelog
 * CSS fix: room-info bug on hover after room description loaded [thierrytiti]
 * I18N: Autodetection of User Locale if no i18n setting is set. [thierrytiti]
 * CSS: Fonts Path: editabable $font-path via sass/variables.scss [thierrytiti]
+* MUC: missing toggle Call handler and function [thierrytiti]
+* Updated documentation about call [thierrytiti]
 
 0.9.3 (2015-05-01)
 ------------------